requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
librouteros==3.2.1
requests==2.31.0
pywhatkit==5.4
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-multipart==0.0.9
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
schedule==1.2.1
